+++ src/ignore.js
@@ -5,6 +5,11 @@ import nodeIgnore from 'ignore';
 
 const LINE_SEPARATOR_REGEX = /(\r|\n|\r\n)/;
 
+/**
+ * Returns a function that checks if a given filename is ignored based on the contents of a file.
+ * @param {string} filename - The path to the file containing the ignore patterns.
+ * @returns {Function} - A function that takes a filename as input and returns true if it is ignored, false otherwise.
+ */
 function getIsIgnored(filename) {
   const ignoreLines = fs
     .readFileSync(filename, 'utf8')
@@ -15,34 +20,25 @@ function getIsIgnored(filename) {
   return instance.ignores.bind(instance);
 }
 
-export function isFilePathMatchedByEslintIgnore(filePath, workspaceDir) {
+/**
+ * Checks if a file path is matched by the ignore rules defined in a specific ignore file.
+ * @param {string} filePath - The path of the file to check.
+ * @param {string} workspaceDirectory - The directory of the workspace.
+ * @param {string} ignoreFileName - The name of the ignore file.
+ * @returns {boolean} - True if the file path is matched by the ignore rules, false otherwise.
+ */
+export default function isFilePathMatchedByIgnore(filePath, workspaceDirectory, ignoreFileName) {
   const options = { cwd: path.dirname(filePath) };
-  if (workspaceDir) {
-    options.stopAt = workspaceDir;
+  if (workspaceDirectory) {
+    options.stopAt = workspaceDirectory;
   }
-  const eslintIgnorePath = findUpSync('.eslintignore', options);
-  if (!eslintIgnorePath) {
+  const ignorePath = findUpSync(ignoreFileName, options);
+  if (!ignorePath) {
     return false;
   }
 
-  const eslintIgnoreDir = path.dirname(eslintIgnorePath);
-  const filePathRelativeToEslintIgnoreDir = path.relative(eslintIgnoreDir, filePath);
-  const isIgnored = getIsIgnored(eslintIgnorePath);
-  return isIgnored(filePathRelativeToEslintIgnoreDir);
-}
-
-export function isFilePathMatchedByPrettierIgnore(filePath, workspaceDir) {
-  const options = { cwd: path.dirname(filePath) };
-  if (workspaceDir) {
-    options.stopAt = workspaceDir;
-  }
-  const prettierIgnorePath = findUpSync('.prettierignore', options);
-  if (!prettierIgnorePath) {
-    return false;
-  }
-
-  const prettierIgnoreDir = path.dirname(prettierIgnorePath);
-  const filePathRelativeToPrettierIgnoreDir = path.relative(prettierIgnoreDir, filePath);
-  const isIgnored = getIsIgnored(prettierIgnorePath);
-  return isIgnored(filePathRelativeToPrettierIgnoreDir);
+  const ignoreDir = path.dirname(ignorePath);
+  const filePathRelativeToIgnoreDir = path.relative(ignoreDir, filePath);
+  const isIgnored = getIsIgnored(ignorePath);
+  return isIgnored(filePathRelativeToIgnoreDir);
 }